added dd_xplatform_utils to house dd.xplatform calls. removed dd.xplatform from filesystem module
