+++ python/tank/__init__.py
@@ -92,6 +92,7 @@ from .commands import list_commands, get_command, SgtkSystemCommand
 from .templatekey import TemplateKey, SequenceKey, IntegerKey, StringKey, TimestampKey
 
 from .dd_utils import dd_jstools_utils
+from .dd_utils import dd_xplatform_utils
 
 # profiling module import
 from . import profiling
