+++ python/tank/util/filesystem.py
@@ -14,7 +14,6 @@ Utility methods for manipulating files and folders
 import os
 import re
 import sys
-import types
 import errno
 import stat
 import shutil
@@ -30,7 +29,6 @@ log = LogManager.get_logger(__name__)
 SKIP_LIST_DEFAULT = [".svn", ".git", ".gitignore", ".hg", ".hgignore"]
 
 # DD
-from dd import xplatform
 from ..dd_utils import dd_jstools_utils
 
 
@@ -493,68 +491,6 @@ def get_unused_path(base_path):
         if not os.path.exists(path):
             break
     return path
-
-
-def build_path(*paths):
-    """Form a path joining multiple `paths` together using the platform
-    dependent path separator. Shell variables contained in `paths` are
-    expanded using the :meth:`os.path.expandvars` method.
-
-    >>> self.build_path('$ABC_PREFIX/python/$PYTHON_VERSION', '$PYTHONPATH')
-    '/repository/abc/python/2.7:/repository/xyz/python:/repository/lol/python/2.7'
-
-    Paths that don't map to an actual directory or file on disk will be
-    excluded from the resulting path.
-
-    """
-    seen = set()
-    resolved = []
-    unresolved = []
-    non_existent = []
-
-    for i, path in enumerate(paths):
-        if not isinstance(path, types.StringTypes):
-            raise TypeError("item #%s is not a string" % i)
-        path = os.path.expandvars(path)
-        for p in xplatform.xsplit(path):
-            if p in seen:
-                continue
-            if "$" in p:  # ignore the ones that were partially expanded
-                unresolved.append(p)
-            elif not os.path.exists(p):
-                non_existent.append(p)
-            else:
-                resolved.append(p)
-            seen.add(p)
-
-    if unresolved or non_existent:
-        if unresolved:
-            log.debug("expandPaths ignored paths with undefined variables:\n- %s", "\n- ".join(unresolved))
-        if non_existent:
-            log.debug("expandPaths ignored paths that don't exist:\n- %s", "\n- ".join(non_existent))
-    return xplatform.xjoin(*resolved)
-
-
-def combine_paths(*paths):
-    """Combines all list arguments by applying a string dot product
-    operation from left to right.
-
-    >>> paths = combinePaths(['shot', 'seq'], ['devl', 'shared'], ['one', 'two'])
-    >>> for path in paths:
-    ...   print path
-    shot/devl/one
-    shot/devl/two
-    shot/shared/one
-    shot/shared/two
-    seq/devl/one
-    seq/devl/two
-    seq/shared/one
-    seq/shared/two
-
-    The catenation behaviour is the same as :func:`os.path.join`.
-    """
-
-    return xplatform.combinePaths(*paths)
 
 
 @with_cleared_umask
